+++ weaver/wps/utils.py
@@ -181,6 +181,31 @@ def get_wps_local_status_location(url_status_location, container, must_exist=Tru
     return out_path
 
 
+def map_wps_output_location(reference, container, reverse=False, exists=True):
+    # type: (str, AnySettingsContainer, bool, bool) -> Optional[str]
+    """
+    Obtains the mapped WPS output location of a file where applicable.
+
+    :param reference: local file path (normal) or file URL (reverse) to be mapped.
+    :param container: retrieve application settings.
+    :param reverse: perform the reverse operation (local path -> URL endpoint), or process normally (URL -> local path).
+    :param exists: ensure that the mapped file exists, otherwise don't map it.
+    :returns: mapped reference that corresponds to the local WPS output location.
+    """
+    settings = get_settings(container)
+    wps_out_dir = get_wps_output_dir(settings)
+    wps_out_url = get_wps_output_url(settings)
+    if reverse and reference.startswith(wps_out_dir):
+        wps_out_ref = reference.replace(wps_out_dir, wps_out_url)
+        if not exists or os.path.isfile(wps_out_ref):
+            return wps_out_ref
+    elif not reverse and reference.startswith(wps_out_url):
+        wps_out_ref = reference.replace(wps_out_url, wps_out_dir)
+        if not exists or os.path.isfile(wps_out_ref):
+            return wps_out_ref
+    return None
+
+
 @cache_region("request")
 def _describe_process_cached(self, identifier, xml=None):
     # type: (WebProcessingService, str, Optional[xml_util.XML]) -> ProcessOWS